+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to the "vsnotes" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.4.0] - 2017-11-14
+### Added
+- Commit and push command added.
+- Commit and push command setting added.
+- Commit and push default commit message added.
+
 ## [0.3.1] - 2017-11-13
 ### Added
 - Icon